Fix lots of compilation warnings

+++ src/Degrains.cpp
@@ -590,12 +590,12 @@ static void VS_CC degrainCreate(const VSMap *in, VSMap *out, [[maybe_unused]] vo
         vectors[0]->ScaleThSCD(d->nSCD1, d->nSCD2, d->vi->format.bitsPerSample);
         double thscdScale = vectors[0]->GetThSCDScaleFactor(d->vi->format.bitsPerSample);
 
-        d->thSAD[0] = d->thSAD[0] * thscdScale + .5;
-        d->thSAD[1] = d->thSAD[1] * thscdScale + .5;
+        d->thSAD[0] = static_cast<int64_t>(d->thSAD[0] * thscdScale + .5);
+        d->thSAD[1] = static_cast<int64_t>(d->thSAD[1] * thscdScale + .5);
         d->thSAD[2] = d->thSAD[1];
 
         if (d->thSAD[0] >= INT_MAX || d->thSAD[1] >= INT_MAX) {
-            int64_t maximum = INT_MAX * thscdScale + .5;
+            int64_t maximum = static_cast<int64_t>(INT_MAX * thscdScale + .5);
 
             bool c = d->thSAD[0] < INT_MAX;
